move msm to its own module
