+++ src/math/mod.rs
@@ -1,3 +1,4 @@
 pub mod field_element;
 pub mod group;
+pub mod msm;
 pub mod polynomial;
